added postTodo() into addTodo()
