+++ src/App.jsx
@@ -7,7 +7,47 @@ function App() {
 	const [todoList, setTodoList] = useState([]);
 	const [isLoading, setIsLoading] = useState('true');
 
+	const postTodo = async (newTodo) => {
+		try {
+			const airtableData = {
+				records: [
+					{
+						fields: {
+							title: `${newTodo}`,
+						},
+					},
+				],
+			};
+
+			const response = await fetch(
+				`https://api.airtable.com/v0/${
+					import.meta.env.VITE_AIRTABLE_BASE_ID
+				}/Default`,
+				{
+					method: 'POST',
+					headers: {
+						'Content-Type': 'application/json',
+						Authorization: `Bearer ${
+							import.meta.env.VITE_AIRTABLE_API_TOKEN
+						}`,
+					},
+					body: JSON.stringify(airtableData),
+				}
+			);
+			if (!response.ok) {
+				const message = `Error has ocurred: ${response.status}`;
+				throw new Error(message);
+			}
+			const dataResponse = await response.json();
+			return dataResponse;
+		} catch (error) {
+			console.log(error.message);
+		}
+	};
+
 	function addTodo(newTodo) {
+		postTodo(newTodo.title);
+
 		setTodoList([...todoList, newTodo]);
 	}
